Added Video
I added the youtube video section. Since there was a new section I changed the listen section (#2) to a gray background so the odd and even columns alternated correctly. I added css to keep video aspect ratio on responsive as well as fixed the icons for listen (it was a camera) and added icon for video.